#4944 - Separated classes into own files
