+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -232,7 +232,9 @@
     <Compile Include="PublishedModelFactoryExtensions.cs" />
     <Compile Include="Services\PropertyValidationService.cs" />
     <Compile Include="Composing\TypeCollectionBuilderBase.cs" />
+    <Compile Include="Sync\DatabaseServerMessengerSyncRepository.cs" />
     <Compile Include="Sync\IServerMessengerSyncRepository.cs" />
+    <Compile Include="Sync\LocalTempFileServerMessengerSyncRepository.cs" />
     <Compile Include="TypeLoaderExtensions.cs" />
     <Compile Include="Composing\WeightAttribute.cs" />
     <Compile Include="Composing\WeightedCollectionBuilderBase.cs" />
